Move dao/entity/repo to channelConfig package
