+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/builder/RepositoryFactory.kt
@@ -2,11 +2,11 @@ package io.getstream.chat.android.livedata.repository.builder
 
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.livedata.repository.ChannelConfigRepository
-import io.getstream.chat.android.livedata.repository.ChannelConfigRepositoryImpl
 import io.getstream.chat.android.livedata.repository.ChannelRepository
 import io.getstream.chat.android.livedata.repository.ChannelRepositoryImpl
 import io.getstream.chat.android.livedata.repository.database.ChatDatabase
+import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigRepository
+import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigRepositoryImpl
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepository
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepositoryImpl
 import io.getstream.chat.android.livedata.repository.domain.queryChannels.QueryChannelsRepository
