+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/database/ChatDatabase.kt
@@ -6,11 +6,8 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import androidx.sqlite.db.SupportSQLiteDatabase
-import io.getstream.chat.android.livedata.dao.ChannelConfigDao
 import io.getstream.chat.android.livedata.dao.ChannelDao
-import io.getstream.chat.android.livedata.entity.ChannelConfigInnerEntity
 import io.getstream.chat.android.livedata.entity.ChannelEntity
-import io.getstream.chat.android.livedata.entity.CommandInnerEntity
 import io.getstream.chat.android.livedata.repository.database.converter.DateConverter
 import io.getstream.chat.android.livedata.repository.database.converter.ExtraDataConverter
 import io.getstream.chat.android.livedata.repository.database.converter.FilterObjectConverter
@@ -18,6 +15,9 @@ import io.getstream.chat.android.livedata.repository.database.converter.ListConv
 import io.getstream.chat.android.livedata.repository.database.converter.MapConverter
 import io.getstream.chat.android.livedata.repository.database.converter.SetConverter
 import io.getstream.chat.android.livedata.repository.database.converter.SyncStatusConverter
+import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigDao
+import io.getstream.chat.android.livedata.repository.domain.channelconfig.ChannelConfigInnerEntity
+import io.getstream.chat.android.livedata.repository.domain.channelconfig.CommandInnerEntity
 import io.getstream.chat.android.livedata.repository.domain.message.MessageDao
 import io.getstream.chat.android.livedata.repository.domain.message.MessageInnerEntity
 import io.getstream.chat.android.livedata.repository.domain.message.attachment.AttachmentEntity
