+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/domain/channelconfig/ChannelConfigEntity.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.livedata.entity
+package io.getstream.chat.android.livedata.repository.domain.channelconfig
 
 import androidx.room.ColumnInfo
 import androidx.room.Embedded
